requirements.txt
Django==4.0.10
djangorestframework==3.13.1
#scikit-learn
pandas